Changed regex, filtered results
Changed the regex to be a bit more comprehensive to account for varied sentence structures, the main change is how we parse the results by filtering through, and returning the raw text in a series of keyword phrases. 

The next step is to find overlap between these and commonly used keywords and entities to find the search terms.